requirements.txt
acres==0.5.0
aiocache==0.12.3
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
av==15.1.0
backoff==2.2.1
bcrypt==4.3.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chroma-bullet==2.2.0
chroma-migrate==0.0.7
chromadb==1.0.21
ci-info==0.3.0
click==8.2.1
clickhouse-connect==0.6.6
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==4.4.2
distlib==0.3.8
distro==1.9.0
duckdb==0.7.1
durationpy==0.10
etelemetry==0.3.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
filelock==3.13.1
fitz==0.0.1.dev2
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.9.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.1
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.0
ipython==8.37.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-groq==0.3.7
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.26
llvmlite==0.44.0
looseversion==1.3.0
lxml==6.0.1
lz4==4.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.7
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.3.2
parso==0.8.4
pathlib==1.0.1
pillow==11.3.0
platformdirs==4.2.0
posthog==5.4.0
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.1
prov==2.1.1
psutil==7.0.0
pure_eval==0.2.3
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.4
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.0.1
rdflib==7.1.4
referencing==0.36.2
regex==2022.9.13
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scipy==1.15.3
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
srt==3.5.3
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tinydb==4.8.2
tokenizers==0.22.0
tomli==2.2.1
torchaudio==2.8.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.25.1
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
